ADSR and Note implementation

+++ Software/Firmware/Core/Inc/audio_core/note/adsr.h
@@ -19,10 +19,10 @@
  */
 typedef struct
 {
-  double attack;        /*!<the time for initial run-up of level from nil to peak in seconds */
-  double decay;         /*!<the time for the subsequent run down from the attack level to the designated sustain level in seconds */
-  double sustain;       /*!<the level during the main sequence of the sound's duration, until the note is off (range from 0 to 1)*/
-  double release;       /*!<the time taken for the level to decay from the sustain level to zero after the note is off in seconds*/
+  float attack;        /*!<the time for initial run-up of level from nil to peak in seconds */
+  float decay;         /*!<the time for the subsequent run down from the attack level to the designated sustain level in seconds */
+  float sustain;       /*!<the level during the main sequence of the sound's duration, until the note is off (range from 0 to 1)*/
+  float release;       /*!<the time taken for the level to decay from the sustain level to zero after the note is off in seconds*/
 } Envelope;
 
 #endif //AUDIO_SYNTH_ADSR_H
